requirements.txt
# NanoBrain Framework v2.0 Dependencies

# Core dependencies
pydantic>=2.0.0
pyyaml>=6.0
asyncio-mqtt>=0.13.0
cachetools>=5.0.0  # For TTL cache in cache_manager

# Optional dependencies for enhanced functionality
jsonschema>=4.0.0  # For YAML schema validation
parsl>=1.3.0       # For HPC execution (optional)
aiofiles>=23.0.0   # For async file operations

# Bioinformatics dependencies
pandas>=2.0.0      # For CSV data processing
fuzzywuzzy>=0.18.0 # For fuzzy string matching in virus name resolution

# Development dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.0.0

# LLM integration (choose one or more)
openai>=1.0.0      # For OpenAI models
anthropic>=0.3.0   # For Anthropic models
langchain>=0.1.0   # For LangChain integration (optional)

# Logging and monitoring
structlog>=23.0.0  # For structured logging (optional)

# Web interface (optional)
fastapi>=0.100.0   # For REST API
uvicorn>=0.20.0    # For ASGI server
streamlit>=1.28.0  # For web UI (optional) 